### 2024-10-20 #7382 @editorialbot
[PRE REVIEW]: PyRASA - Spectral parameterization in python based on IRASA
**Submitting author:** @schmidtfa (<a href="http://orcid.org/0000-0002-9839-1614">Fabian Schmidt</a>)
**Repository:** https://github.com/schmidtfa/pyrasa
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
