### 2024-11-05 05fdac7
added dois to paper.bib
