+++ paper/paper.bib
@@ -1,6 +1,7 @@
 @article{cole2019neurodsp,
   title={NeuroDSP: A package for neural digital signal processing},
   author={Cole, Scott and Donoghue, Thomas and Gao, Richard and Voytek, Bradley},
+  doi = {10.21105/joss.01272},
   journal={Journal of Open Source Software},
   volume={4},
   number={36},
@@ -12,6 +13,7 @@
 @article{donoghue2020parameterizing,
   title={Parameterizing neural power spectra into periodic and aperiodic components},
   author={Donoghue, Thomas and Haller, Matar and Peterson, Erik J and Varma, Paroma and Sebastian, Priyadarshini and Gao, Richard and Noto, Torben and Lara, Antonio H and Wallis, Joni D and Knight, Robert T and others},
+  doi = {10.1038/s41593-020-00744-x},
   journal={Nature neuroscience},
   volume={23},
   number={12},
@@ -23,6 +25,7 @@
 @article{gerster2022separating,
   title={Separating neural oscillations from aperiodic 1/f activity: challenges and recommendations},
   author={Gerster, Moritz and Waterstraat, Gunnar and Litvak, Vladimir and Lehnertz, Klaus and Schnitzler, Alfons and Florin, Esther and Curio, Gabriel and Nikulin, Vadim},
+  doi = {10.1007/s12021-022-09581-8},
   journal={Neuroinformatics},
   volume={20},
   number={4},
@@ -34,6 +37,7 @@
 @article{gramfort2014mne,
   title={MNE software for processing MEG and EEG data},
   author={Gramfort, Alexandre and Luessi, Martin and Larson, Eric and Engemann, Denis A and Strohmeier, Daniel and Brodbeck, Christian and Parkkonen, Lauri and H{\"a}m{\"a}l{\"a}inen, Matti S},
+  doi = {10.1016/j.neuroimage.2013.10.027},
   journal={neuroimage},
   volume={86},
   pages={446--460},
@@ -44,6 +48,7 @@
 @article{harris2020array,
   title={Array programming with NumPy},
   author={Harris, Charles R and Millman, K Jarrod and Van Der Walt, St{\'e}fan J and Gommers, Ralf and Virtanen, Pauli and Cournapeau, David and Wieser, Eric and Taylor, Julian and Berg, Sebastian and Smith, Nathaniel J and others},
+  doi = {10.1038/s41586-020-2649-2},
   journal={Nature},
   volume={585},
   number={7825},
@@ -55,6 +60,7 @@
 @article{oostenveld2011fieldtrip,
   title={FieldTrip: open source software for advanced analysis of MEG, EEG, and invasive electrophysiological data},
   author={Oostenveld, Robert and Fries, Pascal and Maris, Eric and Schoffelen, Jan-Mathijs},
+  doi = {10.1155/2011/156869},
   journal={Computational intelligence and neuroscience},
   volume={2011},
   number={1},
@@ -66,6 +72,7 @@
 @article{wen2016separating,
   title={Separating fractal and oscillatory components in the power spectrum of neurophysiological signal},
   author={Wen, Haiguang and Liu, Zhongming},
+  doi = {10.1007/s10548-015-0448-0},
   journal={Brain topography},
   volume={29},
   pages={13--26},
@@ -76,6 +83,7 @@
 @article{vallat2021open,
   title={An open-source, high-performance tool for automated sleep staging},
   author={Vallat, Raphael and Walker, Matthew P},
+  doi = {10.7554/eLife.70092},
   journal={Elife},
   volume={10},
   pages={e70092},
