### 2024-11-05 #7382 @schmidtfa
fixed the missing dois. so @editorialbot generate pdf
