### 2024-11-05 #7382 @schmidtfa
Also I think Tom Donoghue would be a great fit for a reviewer as he is maintaining fooof/specparam the most widely used package for spectral parametrization.
