### 2025-01-10 #7382 @csoneson
👋🏻 @TomDonoghue, @voytek, @vferat - would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?

>  PyRASA - Spectral parameterization in python based on IRASA
https://github.com/openjournals/joss-reviews/issues/7382

More information about our checklist-based review process can be found [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
Thanks in advance!
