### 2025-01-22 #7382 @mscaudill
Hi @csoneson <https://github.com/csoneson> Thank you for the invitation. I
don't have the bandwidth at this time but I will in early March. If you
don't find reviewers before then, please ping me again.

On Mon, Jan 20, 2025 at 3:13 AM Charlotte Soneson ***@***.***>
wrote:

> 👋🏻 @TomDonoghue <https://github.com/TomDonoghue>, @vferat
> <https://github.com/vferat>, @tuliofalmeida
> <https://github.com/tuliofalmeida>, @mscaudill
> <https://github.com/mscaudill> - would you be interested in reviewing
> this submission for the Journal of Open Source Software (JOSS)?
>
> PyRASA - Spectral parameterization in python based on IRASA
> #7382 <https://github.com/openjournals/joss-reviews/issues/7382>
>
> More information about our checklist-based review process can be found
> here <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>.
> Thanks in advance!
>
