### 2025-01-28 #7382 @LegrandNico
> 👋🏻 [@vferat](https://github.com/vferat), [@axel-loewe](https://github.com/axel-loewe), [@LegrandNico](https://github.com/LegrandNico), [@adswa](https://github.com/adswa) - would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?
>
> > PyRASA - Spectral parameterization in python based on IRASA
> > [#7382](https://github.com/openjournals/joss-reviews/issues/7382)
>
> More information about our checklist-based review process can be found [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Thanks in advance!

Hi @csoneson , sorry I won't have time to review manuscripts this semester. All the best,
