### 2025-02-17 86163a9
Bump prefix-dev/setup-pixi from 0.8.1 to 0.8.2 in the dependencies group
Bumps the dependencies group with 1 update: [prefix-dev/setup-pixi](https://github.com/prefix-dev/setup-pixi).


Updates `prefix-dev/setup-pixi` from 0.8.1 to 0.8.2
- [Release notes](https://github.com/prefix-dev/setup-pixi/releases)
- [Commits](https://github.com/prefix-dev/setup-pixi/compare/v0.8.1...v0.8.2)

---
updated-dependencies:
- dependency-name: prefix-dev/setup-pixi
  dependency-type: direct:production
  update-type: version-update:semver-patch
  dependency-group: dependencies
...

Signed-off-by: dependabot[bot
[…]
