### 2025-02-25 #7382 @OleBialas
> 👋🏻 [@behinger](https://github.com/behinger), [@wmvanvliet](https://github.com/wmvanvliet), [@OleBialas](https://github.com/OleBialas) - would you be interested in reviewing this submission for the Journal of Open Source Software (JOSS)?
>
> > PyRASA - Spectral parameterization in python based on IRASA
> > [#7382](https://github.com/openjournals/joss-reviews/issues/7382)
>
> More information about our checklist-based review process can be found [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Thanks in advance!

This looks very interesting, I'm happy to do the review!
