### 2025-02-25 #7852 @editorialbot
[REVIEW]: PyRASA - Spectral parameterization in python based on IRASA
**Submitting author:** @schmidtfa (<a href="http://orcid.org/0000-0002-9839-1614">Fabian Schmidt</a>)
**Repository:** https://github.com/schmidtfa/pyrasa
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.1
