### 2025-02-26 #7852 @schmidtfa
Hi @wmvanvliet , thanks for taking this on as a reviewer! So Nathans role was mainly conceptual. He was involved in a lot of the discussions that ended up shaping the design and functionality of pyrasa :)
