### 2025-02-26 #7852 @wmvanvliet
Ok, so my first experiences with this package have been very positive. Especially the documentation stands out. I love that the tutorials demonstrate not only how to perform the base case of applying IRANA to a PSD, but also in detail how the algorithm works and common pitfalls. This made it very easy to get started, and I had an easy time comparing `pyrana` to `neurodsp` even though I had never used either package before. During this, some documentation issues, questions and ideas for features came up which I've annotated in the checklist above.
