### 2025-03-12 #7852 @OleBialas
Hi @schmidtfa, and sorry for being a little late.
Like @wmvanvliet, my first impression of the software is really good.
I especially like the documentation and tutorials and their emphasis on correct usage and pitfalls.
I hope I'll have a bit more time over the next days to go more in depth.
