### 2025-04-07 c30f118
Bump prefix-dev/setup-pixi from 0.8.3 to 0.8.4 in the dependencies group
Bumps the dependencies group with 1 update: [prefix-dev/setup-pixi](https://github.com/prefix-dev/setup-pixi).


Updates `prefix-dev/setup-pixi` from 0.8.3 to 0.8.4
- [Release notes](https://github.com/prefix-dev/setup-pixi/releases)
- [Commits](https://github.com/prefix-dev/setup-pixi/compare/v0.8.3...v0.8.4)

---
updated-dependencies:
- dependency-name: prefix-dev/setup-pixi
  dependency-version: 0.8.4
  dependency-type: direct:production
  update-type: version-update:semver-patch
  dependency-group: dependencies
...

S
[…]
