### 2025-04-07 5794aad
add readme2example link
