### 2025-04-08 2d52700
added documentation to irasa_spectrum and tf_spectrum
