### 2025-04-09 c626e2b
added __str__ method and fixed raw_spectrum cutting
