### 2025-04-09 d399e74
speed up testing time and added first plot
