### 2025-04-11 ed11162
added tf_spectrum plot
