+++ tests/test_irasa_sprint.py
@@ -24,6 +24,8 @@ def test_irasa_sprint(ts4sprint, fs, exponent_1, exponent_2):
     )
     irasa_tf.__str__()
 
+    irasa_tf.plot()
+
     # check basic aperiodic detection
     slope_fit = irasa_tf.fit_aperiodic_model(fit_func='fixed')
     #       irasa_tf.aperiodic[np.newaxis, :, :], freqs=irasa_tf.freqs, times=irasa_tf.time,
