+++ pyrasa/utils/irasa_tf_spectrum.py
@@ -1,5 +1,6 @@
 """Output Class of pyrasa.irasa_sprint"""
 
+import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
 from attrs import define
@@ -323,3 +324,138 @@ class IrasaTfSpectrum:
             aperiodic_errors_ch.append(np.array(aperiodic_errors_t).T)
 
         return np.array(aperiodic_errors_ch)
+
+    def plot(
+        self,
+        cur_ch: None | str | int = None,
+        freq_range: tuple[float, float] | None = None,
+        time_range: tuple[float, float] | None = None,
+        log_x: bool = True,
+        vmin: float = 0,
+        vmax: float = 0.1,  # noqa
+    ) -> None:
+        """
+        This function plots a raw power spectrum alongside the seperated periodic and aperiodic spectrum.
+        The main use of this function is for quality control of the IRASA settings that were used to
+        separate periodic and aperiodic spectra.
+
+        NOTE: The y-axis of the raw and aperiodic spectrum is plotted in log-scale
+        and the y-axis of the periodic spectrum is always plotted in linear scale. The reason for this decision is
+        that the periodic spectrum is the result from subtracting the aperiodic spectrum from the raw powerspectrum.
+        This can induce negligible values below 0, very close to 0 or at 0 which creates a weird looking spectrum
+        as logging can result in nans, negative values or infs. The main purpose of the
+        plotting functionality is to allow you to visually inspect whether something went wrong during
+        the IRASA procedure and we feel this is best accomplished using the herein specified settings.
+
+        Parameters
+        ----------
+        freq_range: tuple of (float, float) or None, optional
+            Tuple specifying the frequency range (lower_bound, upper_bound) to which the spectrum should be plotted.
+            If None the full frequency range returned from IRASA is plotted. Default is None.
+        time_range: tuple of (float, float) or None, optional
+            Tuple specifying the selected time range (lower_bound, upper_bound) for which the spectra should be plotted.
+            If None the full time range returned from IRASA is plotted. Default is None.
+        log_x: bool, optional
+            Whether or not to plot the x-axis (Frequencies) log-scaled
+        cur_ch: None or int, optional
+            The channel index to plot. Per default this is set to None which produces a s
+            spectrogram averaged across channels
+        """
+
+        upper_rows = [
+            'raw',
+            'raw',
+            'aperiodic',
+            'aperiodic',
+            'periodic',
+            'periodic',
+        ]
+        lower_rows = [
+            'aperiodic spectrum',
+            'aperiodic spectrum',
+            'aperiodic spectrum',
+            'periodic spectrum',
+            'periodic spectrum',
+            'periodic spectrum',
+        ]
+
+        f, axes = plt.subplot_mosaic(
+            [upper_rows, lower_rows],  # type: ignore
+            layout='tight',
+            width_ratios=[1.0, 1.0, 1.0, 1.0, 1.0, 1.0],
+            height_ratios=[0.5, 1.0],
+            figsize=(8, 6),
+        )
+
+        times, freqs = self.time, self.freqs
+
+        cur_ch = 0
+
+        if freq_range is not None:
+            freq_range_mask = np.logical_and(freqs > freq_range[0], freqs < freq_range[1])
+        else:
+            freq_range_mask = np.ones_like(freqs) == 1
+
+        if time_range is not None:
+            time_range_mask = np.logical_and(times > time_range[0], times < time_range[1])
+        else:
+            time_range_mask = np.ones_like(times) == 1
+
+        sgramms2plot = [
[…]
