### 2025-04-11 aeb1ce4
added additional f_range 2 test
