+++ tests/settings.py
@@ -3,7 +3,7 @@
 import numpy as np
 
 N_SECONDS = 60
-FS = [500]
+FS = [500, 750]
 
 OSC_FREQ = [5, 10, 20]
 MANY_OSC_FREQ = np.arange(2, 30, 2)
