### 2025-04-11 ad43905
added additional f_range 2 test
