### 2025-04-11 ca43b35
added additional f_range 2 test works locally
