+++ pyrasa/irasa.py
@@ -47,8 +47,16 @@ def irasa(
     Computes the aperiodic and periodic components of the power spectrum from a time series using the
     Irregular Resampling Autocorrelation (IRASA) algorithm.
 
-    The IRASA algorithm allows for the decomposition of neural signals into fractal (aperiodic) and
-    oscillatory (periodic) components, providing insight into the underlying dynamics of the data.
+    The IRASA algorithm isolates oscillatory (periodic) activity from fractal (aperiodic) structure in
+    time series data by leveraging the fact that periodic components shift in frequency under resampling,
+    while aperiodic components remain largely invariant.
+
+    IRASA works by systematically resampling the signal using a set of non-integer scaling factors (h and 1/h),
+    computing power spectra for each resampled version, and then computing the geometric mean.
+    Finally the median across resampling factors is computed. This process effectively attenuates spectral peaks
+    caused by oscillations—since they shift inconsistently with h
+    while preserving the broadband, scale-free structure of the aperiodic component. Subtracting this estimate
+    from the original spectrum reveals the periodic component.
 
     Parameters
     ----------
@@ -217,10 +225,23 @@ def irasa_sprint(  # noqa PLR0915 C901
 ) -> IrasaTfSpectrum:
     """
     Computes time-resolved aperiodic and periodic components of the power spectrum from a time series
-    using the Irregular Resampling Autocorrelation (IRASA) algorithm.
+    using the Irregular Resampling Auto-Spectral Analysis (IRASA) algorithm.
 
-    This function is useful for analyzing how the aperiodic and periodic components of the power spectrum
-    change over time, providing a time-frequency decomposition of the signal.
+    This implementation extends the core IRASA method into the time-frequency domain, enabling the
+    decomposition of power spectra into fractal (aperiodic) and oscillatory (periodic) components
+    across time. This is particularly useful for studying non-stationary (neural) signals where spectral
+    dynamics evolve rapidly (e.g., during cognitive tasks or state transitions).
+
+    The IRASA algorithm operates by resampling the signal using non-integer factors (h and 1/h),
+    which causes narrowband oscillatory (periodic) peaks to shift in frequency, while leaving the
+    aperiodic structure largely unaffected. By averaging the resulting power spectra across resampling
+    pairs and computing the median across those averages, IRASA suppresses periodic components and
+    isolates the aperiodic spectrum. Subtracting the aperiodic spectrum from the original spectrum
+    recovers the periodic activity.
+
+    In this time-resolved variant, IRASA is applied to overlapping short-time windows.
+    This allows the algorithm to produce a full time-frequency representation of both the aperiodic
+    and periodic components, providing information in they both change over time.
 
     Parameters
     ----------
