### 2025-04-15 67a351e
improved docstring
