+++ paper/paper.md
@@ -33,16 +33,16 @@ bibliography: paper.bib
 ---
 
 # Summary
-The electric signals generated by physiological activity exhibit both activity patterns that are regularly repeating over time (i.e. periodic) and activity patterns that are temporally irregular (i.e. aperiodic). In recent years several algorithms have been proposed to separate the periodic from the aperiodic parts of the signal, such as the irregular-resampling auto-spectral analysis (IRASA; [@wen2016separating]). IRASA separates periodic and aperiodic components by up-/downsampling time domain signals and computing their respective auto-power spectra. Finally, the aperiodic component is isolated by averaging over the resampled auto-power spectra removing any frequency specific activity. The aperiodic component can then be subtracted from the original power spectrum yielding the residual periodic component. 
-`PyRASA` is a package that is built upon and extends the IRASA algorithm [@wen2016separating]. The package allows its users not only to separate power spectra, but also contains functionality to further parametrize the periodic and aperiodic spectra, by means of peak detection and several slope fitting options (eg. spectral knees). Furthermore, we implemented a function to use the IRASA algorithm in the time-frequency domain allowing for a time-resolved spectral parameterization using IRASA.
+The electric signals generated by physiological activity exhibit both activity patterns that are regularly repeating over time (i.e. periodic) and activity patterns that are temporally irregular (i.e. aperiodic). In recent years several algorithms have been proposed to separate the periodic from the aperiodic parts of the signal, such as the irregular-resampling auto-spectral analysis (IRASA, [@wen2016separating]). IRASA separates periodic and aperiodic components by up-/downsampling time domain signals and computing their respective auto-power spectra. Finally, the aperiodic component is isolated by averaging over the resampled auto-power spectra removing any frequency-specific activity. The aperiodic component can then be subtracted from the original power spectrum yielding the residual periodic component. 
+`PyRASA` is a package that is built upon and extends the IRASA algorithm [@wen2016separating]. The package not only allows its users to separate power spectra, but also contains functionality to further parametrize the periodic and aperiodic spectra, by means of peak detection and several slope fitting options (eg. spectral knees). Furthermore, we implemented a function to use the IRASA algorithm in the time-frequency domain allowing for a time-resolved spectral parameterization using IRASA.
 
 # Statement of Need
[…]
