### 2025-04-28 #7852 @csoneson
Hi @schmidtfa - I have gone through the paper and made a [PR](https://github.com/schmidtfa/pyrasa/pull/90) with a few small suggestions. Please have a look and merge if you agree. In addition, a couple of comment:

- you sometimes use 'parametrization'/'parametrize', and sometimes 'parameterization'/'parameterize'. I would suggest selecting one format and sticking to that.
- on line 47, what does 'herein' refer to? From the sentence, it sounds like it is discussed in the present paper, which it is not - should this be read '...also discussed by Gerster et al (2022)'?
