### 2025-04-28 #7852 @schmidtfa
Thanks for checking the paper @csoneson ! I merged your edits, streamlined my usage of the word parametrization and also fixed the last sentence referring to Gerster et al. 2022 :)
