### 2025-04-28 #7852 @csoneson
Thanks - I missed one small thing: on line 46, "followed a" should be "followed by a". On the same line, I think you could omit the brackets around `@gerster2022separating` to make it appear as Gerster et al. (2022) instead of (Gerster et al., 2022).

Could you make these small changes and generate a new proof with `@editorialbot generate pdf`?

I will make an additional post below outlining the next steps after that.
