### 2025-04-28 938e287
some text corrections
