+++ paper/paper.md
@@ -40,7 +40,7 @@ The electric signals generated by physiological activity exhibit both activity p
 `PyRASA` is an open-source Python package for the parametrization of (neural) power spectra. `PyRASA` has a lightweight architecture that allows users to directly apply the respective functions to separate power spectra to numpy arrays containing time series data [@harris2020array]. However, `PyRASA` can also be optionally extended with functionality to be used in conjunction with `MNE Python` (a popular beginner-friendly tool for the analysis of electrophysiological data, [@gramfort2014mne]), thus offering both beginners in (neural) time series analysis and more advanced users a tool to easily analyze their data. The IRASA algorithm per se has been implemented in a couple of other software packages [@cole2019neurodsp; @vallat2021open; @oostenveld2011fieldtrip], but these implementations of IRASA largely lack functionality to further parametrize periodic and aperiodic spectra in their respective components. We close this gap by offering such functionality for both periodic and aperiodic spectra. For periodic spectra users can extract peak height, bandwidth and center frequency of putative oscillations. Aperiodic spectra can be further analyzed by means of several slope fitting options that allow not only for the assessment of Goodness of fit by several metrics (R2, mean squared error), but also for model comparison using information criteria (BIC/AIC). Additionally, users can easily implement their own custom functions to model aperiodic activity. Furthermore, we implemented a function to use the IRASA algorithm in the time-frequency domain, by computing IRASA over up-/downsampled versions of spectrograms instead of power spectra thereby also allowing for a time-resolved spectral parametrization of (neural) time series data. 
 
 # Related Projects
-`PyRASA’s` functionality is inspired by specparam (formerly `FOOOF`, [@donoghue2020parameterizing]), a popular tool for spectral parametrization built upon a different algorithm that separates power spectra by first flattening the spectrum and then sequentially modelling peaks as gaussians followed a final fit of the aperiodic component. Each algorithm (IRASA vs. specparam) comes with their specific advantages and disadvantages that are also discussed by [@gerster2022separating].
+`PyRASA’s` functionality is inspired by specparam (formerly `FOOOF`, [@donoghue2020parameterizing]), a popular tool for spectral parametrization built upon a different algorithm that separates power spectra by first flattening the spectrum and then sequentially modelling peaks as gaussians followed by a final fit of the aperiodic component. Each algorithm (IRASA vs. specparam) comes with their specific advantages and disadvantages that are also discussed by @gerster2022separating .
 
 The IRASA algorithm has also been implemented as part of software packages `NeuroDSP` [@cole2019neurodsp], `YASA` [@vallat2021open] and `FieldTrip` [@oostenveld2011fieldtrip].
 
