### 2025-04-28 a0095f4
some text corrections
