+++ README.md
@@ -117,4 +117,4 @@ Wen, H., & Liu, Z. (2016). Separating fractal and oscillatory components in the 
 
 If you are using PyRASA it would be nice, if you could additionally cite us (whenever the paper is finally ready):
 
-Schmidt F., Hartmann T., & Weisz, N. (2025). PyRASA - Spectral parametrization in python based on IRASA. SOME JOURNAL THAT LIKES US
+Schmidt F., Hartmann T., & Weisz, N. (2025). PyRASA - Spectral parametrization in python based on IRASA.
