+++ pyrasa/__version__.py
@@ -1,3 +1,3 @@
 """Version information for Pyrasa."""
 
-__version__ = '1.1.1'
+__version__ = '1.1.1.dev0'
