### 2025-05-02 #7852 @schmidtfa
Hi @csoneson ,

I just created a new release version 1.1.1 and archived it on zenodo under this doi: https://doi.org/10.5281/zenodo.15323086.
The title and author list (including ORCIDs) of the release match those in the JOSS paper. I also checked and the license of the archive is the same as in the repo. :)
