### 2025-05-02 #7852 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
