### 2025-05-02 #7852 @csoneson
🎉 Thanks @schmidtfa! It's great that you would be willing to help out as a reviewer - could you add your information in our [reviewer database](https://reviewers.joss.theoj.org/join), in that way you will be findable when a suitable submission comes in!
