### 2025-05-17 #7852 @Kevin-Mattheus-Moerman
@schmidtfa congratulations on this JOSS publication !!!

@csoneson thanks for editing!

And a special thank you to the reviewers! @wmvanvliet, @OleBialas
